env: reorganized directory structure
