+++ ig/flow/GameFlow.java
@@ -1,6 +1,6 @@
 package ig.flow;
 
-import ig.Game;
+import ig.game.Game;
 
 /**
  * {@code GameFlow} is a class used for creating a new thread
