+++ ig/game/GamePanel.java
@@ -1,4 +1,4 @@
-package ig;
+package ig.game;
 
 import javax.swing.JPanel;
 import java.awt.Color;
