--- ig/shape/Point.java
@@ -1,29 +1,0 @@
-package ig.shape;
-
-public class Point {
-
-    private int x;
-    private int y;
-
-    public Point(int x, int y) {
-        setX(x);
-        setY(y);
-    }
-
-    public void setX(int x) {
-        this.x = x;
-    }
-
-    public int getX() {
-        return this.x;
-    }
-
-    public void setY(int y) {
-        this.y = y;
-    }
-
-    public int getY() {
-        return this.y;
-    }
-
-}
